refac: 분포 중복 코드 추출
