+++ test-web/src/components/Stats/Charts/HeatMap/Sens_Heated_Map.js
@@ -1,6 +1,7 @@
 import ApexCharts from 'react-apexcharts';
 import React, { useEffect, useState } from 'react';
 import { statisticSensoryHeated } from '../../../../API/statistic/statisticSensoryHeated';
+import calculateHeatMapChartSeries from './calculateHeatMapChartSeries';
 
 export default function Sens_Heated_Map({
   startDate,
@@ -43,28 +44,7 @@ export default function Sens_Heated_Map({
     umami: '감칠맛',
   };
 
-  let ChartSeries = prop
-    .map((property) => {
-      const uniqueValues = chartData[property].values;
-      const frequencies = new Array(11).fill(0);
-
-      uniqueValues.forEach((value) => {
-        if (value > 10) {
-          frequencies[10] += 1;
-        } else if (value < 1) {
-          frequencies[0] += 1;
-        } else {
-          const index = Math.floor(value);
-          frequencies[index] += 1;
-        }
-      });
-
-      return {
-        name: y_axis[property] || property,
-        data: frequencies,
-      };
-    })
-    .reverse();
+  const ChartSeries = calculateHeatMapChartSeries(prop, chartData, y_axis);
 
   const ChartOption = {
     chart: {
