+++ test-web/src/components/Stats/Charts/HeatMap/Sens_Proc_Map.js
@@ -1,6 +1,7 @@
 import ApexCharts from 'react-apexcharts';
 import React, { useEffect, useState } from 'react';
 import { statisticSensoryProcessed } from '../../../../API/statistic/statisticSensoryProcessed';
+import calculateHeatMapChartSeries from './calculateHeatMapChartSeries';
 
 export default function Sens_Proc_Map({
   startDate,
@@ -43,32 +44,7 @@ export default function Sens_Proc_Map({
     texture: '조직감',
   };
 
-  let ChartSeries = [];
-  if (prop.length > 0) {
-    ChartSeries = prop
-      .map((property) => {
-        const uniqueValues = chartData[property].values;
-        const frequencies = new Array(11).fill(0);
-
-        uniqueValues.forEach((value) => {
-          if (value > 10) {
-            frequencies[10] += 1;
-          } else if (value < 1) {
-            frequencies[0] += 1;
-          } else {
-            const index = Math.floor(value);
-            frequencies[index] += 1;
-          }
-        });
-
-        return {
-          name: y_axis[property] || property,
-          data: frequencies,
-        };
-      })
-      .reverse();
-  }
-
+  const ChartSeries = calculateHeatMapChartSeries(prop, chartData, y_axis);
   const ChartOption = {
     chart: {
       height: 450,
